+++ client/src/components/LoginForm/LoginForm.styles.tsx
@@ -1,0 +1,20 @@
+import styled from "styled-components";
+
+export const LoginFormContainer = styled.div`
+    width: 100%;
+    padding: 0 40px;
+    div {
+        input {
+            font-size: 1.4rem;
+        }
+    }
+
+    span {
+        color: red;
+        font-size: 1rem;
+    }
+
+    button {
+        margin-top: 25px;
+    }
+`;
